nse_scraper/requirements.txt
selenium
webdriver-manager
requests
pandas
mysql-connector-python
schedule
